fix: warn

+++ synctv-core/src/cache/room_cache.rs
@@ -46,7 +46,7 @@ impl RoomCache {
         key_prefix: String,
     ) -> Result<Self> {
         let l1_cache = moka::future::CacheBuilder::new(l1_max_capacity)
-            .time_to_live(std::time::Duration::from_secs(l1_ttl_minutes as u64 * 60))
+            .time_to_live(std::time::Duration::from_secs(l1_ttl_minutes * 60))
             .build();
 
         Ok(Self {
